removed extra js file for simplicity
