+++ assets/javascript/app.js
@@ -1,11 +1,23 @@
 $(function () {
 
-    const cards = require("./cards");
+    $.getScript("assets/javascript/cards.js", function() {
+        console.log("cards.js imported");
+     });
 
     //elements hidden on page load
     $(".resume-con").hide();
     $(".res-btn").hide();
     $(".back-btn").hide();
+
+    let cards = [
+        {port1: [
+            {title: "Gif Generator"},
+            {para: "This app uses the Giphy API to generate 10 gifs based on whichever button is clicked. The user can also add their own buttons. Built using Javascript, JQuery, CSS, and Bootstrap."},
+            {linkOne: "https://edwardsjtimothy.github.io/Gif-Generator/"},
+            {linkTwo: "https://github.com/edwardsjtimothy/Gif-Generator"}
+        ]},
+    ];
+    
 
     //generates portfolio cards
     const cardGen = () => {
