--- assets/javascript/cards.js
@@ -1,12 +1,0 @@
-let cards = [
-    {port1: [
-        {title: "Gif Generator"},
-        {para: "This app uses the Giphy API to generate 10 gifs based on whichever button is clicked. The user can also add their own buttons. Built using Javascript, JQuery, CSS, and Bootstrap."},
-        {linkOne: "https://edwardsjtimothy.github.io/Gif-Generator/"},
-        {linkTwo: "https://github.com/edwardsjtimothy/Gif-Generator"}
-    ]},
-];
-
-module.exports = {
-    cards: cards,
-};
